Remove old ecc
- remove old ecc
- retain ecc boilderplate
- implement invoke
