+++ ecc/.gitignore
@@ -1,3 +1,1 @@
-# SPDX-License-Identifier: Apache-2.0
 ecc
-tags
