+++ ecc/Dockerfile.boilerplate-ecc
@@ -22,9 +22,6 @@ ENV SGX_MODE=${SGX_MODE}
 # define here a env which makes it easy recognizable which mode
 # the container is. No default, though, as we do not control
 # the build and rely on a proper value provided from outside.
-ARG FPC_ATTESTATION_TYPE
-ENV FPC_ATTESTATION_TYPE=${FPC_ATTESTATION_TYPE}
-
 
 ARG CC_NAME="ecc"
 ARG CC_PATH="/usr/local/bin"
