+++ ecc/chaincode/ecc.go
@@ -13,8 +13,8 @@ import (
 	"github.com/hyperledger/fabric/protoutil"
 
 	"github.com/golang/protobuf/ptypes"
-	"github.com/hyperledger-labs/fabric-private-chaincode/ecc_mock/chaincode/enclave"
-	"github.com/hyperledger-labs/fabric-private-chaincode/ecc_mock/chaincode/ercc"
+	"github.com/hyperledger-labs/fabric-private-chaincode/ecc/chaincode/enclave"
+	"github.com/hyperledger-labs/fabric-private-chaincode/ecc/chaincode/ercc"
 	"github.com/hyperledger-labs/fabric-private-chaincode/internal/protos"
 	"github.com/hyperledger-labs/fabric-private-chaincode/internal/utils"
 	"github.com/hyperledger/fabric-chaincode-go/shim"
@@ -62,34 +62,34 @@ func (t *EnclaveChaincode) initEnclave(stub shim.ChaincodeStubInterface) pb.Resp
 	// fetch cc params and host params
 	chaincodeParams, err := extractChaincodeParams(stub)
 	if err != nil {
-		shim.Error(err.Error())
+		return shim.Error(err.Error())
 	}
 	serializedChaincodeParams, err := protoutil.Marshal(chaincodeParams)
 	if err != nil {
-		shim.Error(err.Error())
+		return shim.Error(err.Error())
 	}
 
 	hostParams, err := extractHostParams(stub)
 	if err != nil {
-		shim.Error(err.Error())
+		return shim.Error(err.Error())
 	}
 	serializedHostParams, err := protoutil.Marshal(hostParams)
 	if err != nil {
-		shim.Error(err.Error())
+		return shim.Error(err.Error())
 	}
 
 	attestationParams, err := extractAttestationParams(stub)
 	if err != nil {
-		shim.Error(err.Error())
+		return shim.Error(err.Error())
 	}
 	serializedAttestationParams, err := protoutil.Marshal(attestationParams)
 	if err != nil {
-		shim.Error(err.Error())
+		return shim.Error(err.Error())
 	}
 
 	credentialsBytes, err := t.enclave.Init(serializedChaincodeParams, serializedHostParams, serializedAttestationParams)
 	if err != nil {
-		shim.Error(err.Error())
+		return shim.Error(err.Error())
 	}
 
 	// return credentials
@@ -103,48 +103,65 @@ func (t *EnclaveChaincode) invoke(stub shim.ChaincodeStubInterface) pb.Response 
 	}
 
 	// call enclave
-	responseBytes, errInvoke := t.enclave.ChaincodeInvoke(stub)
+	var errMsg string
+	b64ChaincodeResponseMessage, errInvoke := t.enclave.ChaincodeInvoke(stub)
 	if errInvoke != nil {
-		errMsg := fmt.Sprintf("t.enclave.ChaincodeInvoke failed: %s", errInvoke)
+		errMsg = fmt.Sprintf("t.enclave.Invoke failed: %s", errInvoke)
 		logger.Errorf(errMsg)
-		// likely a chaincode error, so we still want response go back ...
-		return pb.Response{
+		// likely a chaincode error, so we stil want response go back ...
+	}
+
+	var response pb.Response
+	if errInvoke == nil {
+		response = pb.Response{
+			Status:  shim.OK,
+			Payload: b64ChaincodeResponseMessage,
+			Message: errMsg,
+		}
+	} else {
+		response = pb.Response{
 			Status:  shim.ERROR,
-			Payload: responseBytes,
+			Payload: b64ChaincodeResponseMessage,
 			Message: errMsg,
 		}
 	}
 
-	return shim.Success(responseBytes)
+	return response
 }
 
 func (t *EnclaveChaincode) endorse(stub shim.ChaincodeStubInterface) pb.Response {
 
 	chaincodeParams, err := extractChaincodeParams(stub)
 	if err != nil {
-		shim.Error(err.Error())
+		return shim.Error(err.Error())
 	}
 
 	responseMsg, err := extractChaincodeResponseMessage(stub)
 	if err != nil {
-		shim.Error(err.Error())
+		return shim.Error(err.Error())
 	}
+
+	logger.Infof("try to get credentials from ERCC for channel: %s ccId: %s EnclaveId: %s ", chaincodeParams.ChannelId, chaincodeParams.ChaincodeId, responseMsg.EnclaveId)
 
 	// get corresponding enclave credentials from ercc
 	credentials, err := ercc.QueryEnclaveCredentials(stub, chaincodeParams.ChannelId, chaincodeParams.ChaincodeId, responseMsg.EnclaveId)
 	if err != nil {
-		shim.Error(err.Error())
+		return shim.Error(err.Error())
+	}
+
+	if credentials == nil {
+		return shim.Error(fmt.Sprintf("no credentials found for enclaveId = %s", responseMsg.EnclaveId))
 	}
 
 	var attestedData protos.AttestedData
 	if err := ptypes.UnmarshalAny(credentials.SerializedAttestedData, &attestedData); err != nil {
-		shim.Error(err.Error())
+		return shim.Error(err.Error())
 	}
 
 	// check cc params match credentials
 	// check cc params chaincode def
 	if ccParamsMatch(attestedData.CcParams, chaincodeParams) {
-		shim.Error("ccParams don't match")
+		return shim.Error("ccParams don't match")
 	}
 
 	// check cc param.MSPID matches MSPID of endorser (Post-MVP)
@@ -152,13 +169,13 @@ func (t *EnclaveChaincode) endorse(stub shim.ChaincodeStubInterface) pb.Response
 	// replay read/writes from kvrwset from enclave (to prepare commitment to ledger) and extract kvrwset for subsequent validation
 	readset, writeset, err := utils.ReplayReadWrites(stub, responseMsg.RwSet)
 	if err != nil {
-		shim.Error(err.Error())
+		return shim.Error(err.Error())
 	}
 
 	// validate enclave endorsement signature
 	err = utils.Validate(responseMsg, readset, writeset, &attestedData)
 	if err != nil {
-		shim.Error(err.Error())
+		return shim.Error(err.Error())
 	}
 
 	return shim.Success(nil)
