+++ ecc/chaincode/enclave/ecc-enclave-include
@@ -1,0 +1,1 @@
+../../../ecc_enclave/_build/include
