+++ ecc/chaincode/enclave/ecc-enclave-lib
@@ -1,0 +1,1 @@
+../../../ecc_enclave/_build/lib
