+++ ecc/chaincode/enclave/interface.go
