+++ ecc/chaincode/enclave/logging_cgo.go
