+++ ecc/chaincode/enclave/logging_go.go
