+++ ecc/chaincode/enclave/shim.go
