+++ ecc/chaincode/ercc/stub.go
