+++ ecc/chaincode/utils.go
