--- ecc/enclave/enclave_stub_test.go
@@ -1,133 +1,0 @@
-/*
-Copyright IBM Corp. All Rights Reserved.
-
-SPDX-License-Identifier: Apache-2.0
-*/
-
-package enclave
-
-import (
-	"testing"
-)
-
-const enclaveLibFile = "lib/enclave.signed.so"
-
-/* Disabled as relying on old attestation ...
-
-func TestEnclaveStub_RemoteAttestation(t *testing.T) {
-	stub := NewEnclave()
-	arr := []byte{}
-	_, err := stub.Create(enclaveLibFile, arr, arr, arr)
-	if err != nil {
-		t.Fatalf("Create returned error %s", err)
-	}
-
-	// start without binding
-	stub.Bind(nil, nil)
-
-	mockRegistry := ercc.MockEnclaveRegistryStub{}
-	spid, _ := mockRegistry.GetSPID(nil, "", "")
-
-	//TODO: retrieve sigrl
-	sig_rl := []byte(nil)
-	sig_rl_size := uint(0)
-
-	quoteAsBytes, pkBytes, err := stub.GetRemoteAttestationReport(spid, sig_rl, sig_rl_size)
-	if err != nil {
-		t.Fatalf("Attestation returned error %s", err)
-	}
-	if quoteAsBytes == nil {
-		t.Fatalf("quote is nil")
-	}
-	if pkBytes == nil {
-		t.Fatalf("pkBytes is nil")
-	}
-
-	// check pk is in quote
-	q, err := attestation.QuoteFromBytes(quoteAsBytes)
-	if err != nil {
-		t.Fatalf("Can not parse quote %s", err)
-	}
-
-	t.Logf("pk der:\n%s", hex.Dump(pkBytes))
-
-	pub, err := x509.ParsePKIXPublicKey(pkBytes)
-	if err != nil {
-		t.Fatalf("x509.ParsePKIXPublicKey error %s", err)
-	}
-
-	ecdsaPublickey, ok := pub.(*ecdsa.PublicKey)
-	if !ok {
-		t.Fatalf("enclave key is not ecdsa key")
-	}
-
-	t.Logf("X:\n%s\nY:\n%s", hex.Dump(ecdsaPublickey.X.Bytes()), hex.Dump(ecdsaPublickey.Y.Bytes()))
-
-	// calc pk raw hash which is in quote
-	h := sha256.New()
-	h.Write(ecdsaPublickey.X.Bytes())
-	h.Write(ecdsaPublickey.Y.Bytes())
-	enclavePKRawHash := h.Sum(nil)
-
-	if !reflect.DeepEqual(enclavePKRawHash[:32], q.ReportData[:32]) {
-		t.Fatalf("enclave pk does not match quote")
-	}
-
-	enclavePkHash := sha256.Sum256(pkBytes)
-
-	t.Logf("Enclave pk base64:\n%s", base64.StdEncoding.EncodeToString(pkBytes))
-	t.Logf("Enclave pk hash base64:\n%s", base64.StdEncoding.EncodeToString(enclavePkHash[:]))
-	t.Logf("Quote base64:\n%s", base64.StdEncoding.EncodeToString(quoteAsBytes))
-}
-*/
-
-func TestEnclaveStub_Invoke(t *testing.T) {
-	stub := NewEnclave()
-	arr := []byte{}
-	_, err := stub.Create(enclaveLibFile, arr, arr, arr)
-	if err != nil {
-		t.Fatalf("Create returned error %s", err)
-	}
-
-	// start without binding
-	stub.Bind(nil, nil)
-
-	_, err = stub.Invoke(nil)
-	if err == nil {
-		t.Fatalf("error expected")
-	}
-}
-
-func TestEnclaveStub_GetPublicKey(t *testing.T) {
-	stub := NewEnclave()
-	arr := []byte{}
-	_, err := stub.Create(enclaveLibFile, arr, arr, arr)
-	if err != nil {
-		t.Fatalf("Create returned error %s", err)
-	}
-
-	// start without binding
-	stub.Bind(nil, nil)
-}
-
-func TestEnclaveStub_Create(t *testing.T) {
-	stub := NewEnclave()
-	arr := []byte{}
-	_, err := stub.Create(enclaveLibFile, arr, arr, arr)
-	if err != nil {
-		t.Fatalf("Create returned error %s", err)
-	}
-}
-
-func TestEnclaveStub_Destroy(t *testing.T) {
-	stub := NewEnclave()
-	arr := []byte{}
-	_, err := stub.Create(enclaveLibFile, arr, arr, arr)
-	if err != nil {
-		t.Fatalf("Create returned error %s", err)
-	}
-	err = stub.Destroy()
-	if err != nil {
-		t.Fatalf("Deate returned error %s", err)
-	}
-}
