--- ecc_mock/.gitignore
@@ -1,1 +1,0 @@
-ecc
