--- ecc_mock/Dockerfile
@@ -1,11 +1,0 @@
-# Copyright IBM Corp. All Rights Reserved.
-# Copyright 2020 Intel Corporation
-#
-# SPDX-License-Identifier: Apache-2.0
-
-FROM fpc/ecc_enclave
-
-COPY ecc .
-
-EXPOSE 9999
-CMD ["ecc"]
